+++ src/reporting/scope.js
